[test] Split some integration tests
