+++ internal/venice-test-common/src/integrationTest/java/com/linkedin/venice/endToEnd/PartialUpdateTest.java
@@ -1,15 +1,6 @@
 package com.linkedin.venice.endToEnd;
 
 import static com.linkedin.davinci.stats.HostLevelIngestionStats.ASSEMBLED_RMD_SIZE_IN_BYTES;
-import static com.linkedin.davinci.store.rocksdb.RocksDBServerConfig.ROCKSDB_BLOCK_CACHE_SIZE_IN_BYTES;
-import static com.linkedin.davinci.store.rocksdb.RocksDBServerConfig.ROCKSDB_PLAIN_TABLE_FORMAT_ENABLED;
-import static com.linkedin.venice.ConfigKeys.CLIENT_SYSTEM_STORE_REPOSITORY_REFRESH_INTERVAL_SECONDS;
-import static com.linkedin.venice.ConfigKeys.CLIENT_USE_SYSTEM_STORE_REPOSITORY;
-import static com.linkedin.venice.ConfigKeys.DATA_BASE_PATH;
-import static com.linkedin.venice.ConfigKeys.DAVINCI_PUSH_STATUS_CHECK_INTERVAL_IN_MS;
-import static com.linkedin.venice.ConfigKeys.PERSISTENCE_TYPE;
-import static com.linkedin.venice.endToEnd.TestBatch.TEST_TIMEOUT;
-import static com.linkedin.venice.meta.PersistenceType.ROCKS_DB;
 import static com.linkedin.venice.schema.rmd.RmdConstants.TIMESTAMP_FIELD_NAME;
 import static com.linkedin.venice.schema.rmd.v1.CollectionRmdTimestamp.ACTIVE_ELEM_TS_FIELD_NAME;
 import static com.linkedin.venice.schema.rmd.v1.CollectionRmdTimestamp.DELETED_ELEM_TS_FIELD_NAME;
@@ -19,45 +10,31 @@ import static com.linkedin.venice.utils.IntegrationTestPushUtils.sendStreamingDe
 import static com.linkedin.venice.utils.IntegrationTestPushUtils.sendStreamingRecord;
 import static com.linkedin.venice.utils.IntegrationTestPushUtils.verifyConsumerThreadPoolFor;
 import static com.linkedin.venice.utils.TestUtils.assertCommand;
-import static com.linkedin.venice.utils.TestWriteUtils.NAME_RECORD_V1_SCHEMA;
 import static com.linkedin.venice.utils.TestWriteUtils.NAME_RECORD_V2_SCHEMA;
 import static com.linkedin.venice.utils.TestWriteUtils.STRING_SCHEMA;
 import static com.linkedin.venice.utils.TestWriteUtils.getTempDataDirectory;
 import static com.linkedin.venice.utils.TestWriteUtils.loadFileAsString;
 import static com.linkedin.venice.utils.TestWriteUtils.writeSimpleAvroFileWithStringToNameRecordV1Schema;
-import static com.linkedin.venice.utils.TestWriteUtils.writeSimpleAvroFileWithStringToPartialUpdateOpRecordSchema;
 import static com.linkedin.venice.utils.TestWriteUtils.writeSimpleAvroFileWithStringToUserWithStringMapSchema;
-import static com.linkedin.venice.vpj.VenicePushJobConstants.DATA_WRITER_COMPUTE_JOB_CLASS;
 import static com.linkedin.venice.vpj.VenicePushJobConstants.DEFAULT_KEY_FIELD_PROP;
 import static com.linkedin.venice.vpj.VenicePushJobConstants.DEFAULT_VALUE_FIELD_PROP;
-import static com.linkedin.venice.vpj.VenicePushJobConstants.ENABLE_WRITE_COMPUTE;
-import static com.linkedin.venice.vpj.VenicePushJobConstants.INCREMENTAL_PUSH;
 import static com.linkedin.venice.vpj.VenicePushJobConstants.KAFKA_INPUT_BROKER_URL;
 import static com.linkedin.venice.vpj.VenicePushJobConstants.KAFKA_INPUT_MAX_RECORDS_PER_MAPPER;
-import static com.linkedin.venice.vpj.VenicePushJobConstants.REPUSH_TTL_ENABLE;
-import static com.linkedin.venice.vpj.VenicePushJobConstants.REPUSH_TTL_START_TIMESTAMP;
 import static com.linkedin.venice.vpj.VenicePushJobConstants.REWIND_TIME_IN_SECONDS_OVERRIDE;
 import static com.linkedin.venice.vpj.VenicePushJobConstants.SOURCE_KAFKA;
-import static com.linkedin.venice.vpj.VenicePushJobConstants.SPARK_NATIVE_INPUT_FORMAT_ENABLED;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 import static org.testng.Assert.assertEquals;
 import static org.testng.Assert.assertFalse;
 import static org.testng.Assert.assertNotNull;
 import static org.testng.Assert.assertNull;
-import static org.testng.Assert.assertThrows;
 import static org.testng.Assert.assertTrue;
 
 import com.linkedin.avroutil1.compatibility.AvroCompatibilityHelper;
 import com.linkedin.d2.balancer.D2Client;
 import com.linkedin.d2.balancer.D2ClientBuilder;
-import com.linkedin.davinci.client.DaVinciClient;
-import com.linkedin.davinci.client.DaVinciConfig;
-import com.linkedin.davinci.client.StorageClass;
-import com.linkedin.davinci.client.factory.CachingDaVinciClientFactory;
 import com.linkedin.davinci.kafka.consumer.ConsumerPoolType;
 import com.linkedin.davinci.kafka.consumer.KafkaConsumerServiceDelegator;
-import com.linkedin.davinci.kafka.consumer.StoreIngestionTaskBackdoor;
 import com.linkedin.davinci.replication.RmdWithValueSchemaId;
 import com.linkedin.davinci.replication.merge.RmdSerDe;
 import com.linkedin.davinci.replication.merge.StringAnnotatedStoreSchemaCache;
@@ -73,7 +50,6 @@ import com.linkedin.venice.client.store.ClientFactory;
 import com.linkedin.venice.compression.CompressionStrategy;
 import com.linkedin.venice.controllerapi.ControllerClient;
 import com.linkedin.venice.controllerapi.ControllerResponse;
-import com.linkedin.venice.controllerapi.NewStoreResponse;
 import com.linkedin.venice.controllerapi.SchemaResponse;
 import com.linkedin.venice.controllerapi.UpdateStoreQueryParams;
 import com.linkedin.venice.controllerapi.VersionCreationResponse;
@@ -83,10 +59,8 @@ import com.linkedin.venice.integration.utils.VeniceClusterWrapper;
 import com.linkedin.venice.integration.utils.VeniceControllerWrapper;
 import com.linkedin.venice.integration.utils.VeniceMultiClusterWrapper;
 import com.linkedin.venice.integration.utils.VeniceMultiRegionClusterCreateOptions;
-import com.linkedin.venice.integration.utils.VeniceRouterWrapper;
 import com.linkedin.venice.integration.utils.VeniceServerWrapper;
 import com.linkedin.venice.integration.utils.VeniceTwoLayerMultiRegionMultiClusterWrapper;
-import com.linkedin.venice.meta.BackupStrategy;
 import com.linkedin.venice.meta.ReadOnlySchemaRepository;
 import com.linkedin.venice.meta.Store;
 import com.linkedin.venice.meta.StoreInfo;
@@ -103,22 +77,16 @@ import com.linkedin.venice.schema.writecompute.DerivedSchemaEntry;
 import com.linkedin.venice.schema.writecompute.WriteComputeSchemaConverter;
 import com.linkedin.venice.serialization.avro.AvroProtocolDefinition;
 import com.linkedin.venice.serialization.avro.ChunkedValueManifestSerializer;
-import com.linkedin.venice.spark.datawriter.jobs.DataWriterSparkJob;
 import com.linkedin.venice.stats.AbstractVeniceStats;
-import com.linkedin.venice.stats.VeniceMetricsRepository;
 import com.linkedin.venice.storage.protocol.ChunkedValueManifest;
 import com.linkedin.venice.tehuti.MetricsUtils;
 import com.linkedin.venice.utils.DataProviderUtils;
 import com.linkedin.venice.utils.IntegrationTestPushUtils;
-import com.linkedin.venice.utils.PropertyBuilder;
 import com.linkedin.venice.utils.TestUtils;
-import com.linkedin.venice.utils.TestWriteUtils;
 import com.linkedin.venice.utils.Time;
 import com.linkedin.venice.utils.Utils;
-import com.linkedin.venice.utils.VeniceProperties;
 import com.linkedin.venice.writer.update.UpdateBuilder;
 import com.linkedin.venice.writer.update.UpdateBuilderImpl;
-import io.tehuti.metrics.MetricsRepository;
 import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.io.IOException;
@@ -135,7 +103,6 @@ import java.util.concurrent.TimeUnit;
 import java.util.function.BiConsumer;
 import java.util.function.Consumer;
 import java.util.stream.Collectors;
-import org.apache.avro.AvroRuntimeException;
 import org.apache.avro.Schema;
 import org.apache.avro.generic.GenericData;
 import org.apache.avro.generic.GenericDatumWriter;
@@ -151,7 +118,7 @@ import org.testng.annotations.Test;
 
 
 /**
- * This class includes tests on partial update (Write Compute) with a setup that has both the parent and child controllers.
+ * This class includes tests on partial update core functionality.
  */
 public class PartialUpdateTest {
   private static final int NUMBER_OF_CHILD_DATACENTERS = 2;
@@ -172,10 +139,6 @@ public class PartialUpdateTest {
   private List<VeniceMultiClusterWrapper> childDatacenters;
   private D2Client d2ClientDC0;
 
-  protected boolean isAAWCParallelProcessingEnabled() {
-    return false;
-  }
-
   protected boolean isHeartbeatReadyToServeCheckEnabled() {
     return false;
   }
@@ -187,9 +150,7 @@ public class PartialUpdateTest {
     serverProperties.put(
         ConfigKeys.SERVER_CONSUMER_POOL_ALLOCATION_STRATEGY,
         KafkaConsumerServiceDelegator.ConsumerPoolStrategyType.CURRENT_VERSION_PRIORITIZATION.name());
-    serverProperties.put(
-        ConfigKeys.SERVER_AA_WC_WORKLOAD_PARALLEL_PROCESSING_ENABLED,
-        Boolean.toString(isAAWCParallelProcessingEnabled()));
+    serverProperties.put(ConfigKeys.SERVER_AA_WC_WORKLOAD_PARALLEL_PROCESSING_ENABLED, Boolean.toString(true));
     serverProperties.put(
         ConfigKeys.SERVER_USE_HEARTBEAT_LAG_FOR_READY_TO_SERVE_CHECK_ENABLED,
         Boolean.toString(isHeartbeatReadyToServeCheckEnabled()));
@@ -221,197 +182,6 @@ public class PartialUpdateTest {
         .setZkStartupTimeout(3, TimeUnit.SECONDS)
         .build();
     D2ClientUtils.startClient(d2ClientDC0);
-  }
-
-  @Test(timeOut = TEST_TIMEOUT_MS)
-  public void testRepushWithChunkingFlagChanged() {
-    final String storeName = Utils.getUniqueString("reproduce");
-    String parentControllerUrl = parentController.getControllerUrl();
-    Schema keySchema = AvroCompatibilityHelper.parse(loadFileAsString("UserKey.avsc"));
-    Schema valueSchema = AvroCompatibilityHelper.parse(loadFileAsString("UserValue.avsc"));
-    Schema writeComputeSchema = WriteComputeSchemaConverter.getInstance().convertFromValueRecordSchema(valueSchema);
-
-    try (ControllerClient parentControllerClient = new ControllerClient(CLUSTER_NAME, parentControllerUrl)) {
-      assertCommand(
-          parentControllerClient.createNewStore(storeName, "test_owner", keySchema.toString(), valueSchema.toString()));
-      UpdateStoreQueryParams updateStoreParams =
-          new UpdateStoreQueryParams().setStorageQuotaInByte(Store.UNLIMITED_STORAGE_QUOTA)
-              .setCompressionStrategy(CompressionStrategy.NO_OP)
-              .setWriteComputationEnabled(true)
-              .setHybridRewindSeconds(86400L)
-              .setHybridOffsetLagThreshold(10L);
-      ControllerResponse updateStoreResponse =
-          parentControllerClient.retryableRequest(5, c -> c.updateStore(storeName, updateStoreParams));
-      assertFalse(updateStoreResponse.isError(), "Update store got error: " + updateStoreResponse.getError());
-
-      VersionCreationResponse response = parentControllerClient.emptyPush(storeName, "test_push_id", 1000);
-      assertEquals(response.getVersion(), 1);
-      assertFalse(response.isError(), "Empty push to parent colo should succeed");
-      TestUtils.waitForNonDeterministicPushCompletion(
-          Version.composeKafkaTopic(storeName, 1),
-          parentControllerClient,
-          30,
-          TimeUnit.SECONDS);
-
-      VeniceClusterWrapper veniceCluster = childDatacenters.get(0).getClusters().get(CLUSTER_NAME);
-      SystemProducer veniceProducer = getSamzaProducer(veniceCluster, storeName, Version.PushType.STREAM);
-      GenericRecord keyRecord = new GenericData.Record(keySchema);
-      keyRecord.put("learnerUrn", "urn:li:member:682787898");
-      keyRecord.put("query", "python");
-      GenericRecord checkpointKeyRecord = new GenericData.Record(keySchema);
-      checkpointKeyRecord.put("learnerUrn", "urn:li:member:123");
-      checkpointKeyRecord.put("query", "python");
-
-      GenericRecord partialUpdateRecord = new UpdateBuilderImpl(writeComputeSchema)
-          .setElementsToAddToListField("blockedContentsUrns", Collections.singletonList("urn:li:lyndaCourse:751323"))
-          .build();
-      sendStreamingRecord(veniceProducer, storeName, keyRecord, partialUpdateRecord);
-
-      // Perform one time repush to make sure repush can handle RMD chunks data correctly.
-      Properties props =
-          IntegrationTestPushUtils.defaultVPJProps(multiRegionMultiClusterWrapper, "dummyInputPath", storeName);
-      props.setProperty(SOURCE_KAFKA, "true");
-      props.setProperty(KAFKA_INPUT_BROKER_URL, veniceCluster.getPubSubBrokerWrapper().getAddress());
-      props.setProperty(KAFKA_INPUT_MAX_RECORDS_PER_MAPPER, "5");
-      // intentionally stop re-consuming from RT so stale records don't affect the testing results
-      // props.put(REWIND_TIME_IN_SECONDS_OVERRIDE, 0);
-      IntegrationTestPushUtils.runVPJ(props);
-      TestUtils.waitForNonDeterministicPushCompletion(
-          Version.composeKafkaTopic(storeName, 2),
-          parentControllerClient,
-          30,
-          TimeUnit.SECONDS);
-
-      try (AvroGenericStoreClient<Object, Object> storeReader = ClientFactory.getAndStartGenericAvroClient(
-          ClientConfig.defaultGenericClientConfig(storeName).setVeniceURL(veniceCluster.getRandomRouterURL()))) {
-        TestUtils.waitForNonDeterministicAssertion(10, TimeUnit.SECONDS, true, () -> {
-          try {
-            GenericRecord value = (GenericRecord) storeReader.get(keyRecord).get();
-            assertNotNull(value, "key " + keyRecord + " should not be missing!");
-          } catch (Exception e) {
-            throw new VeniceException(e);
-          }
-        });
-
-        // Enable chunking
-        UpdateStoreQueryParams newUpdateStoreParams =
-            new UpdateStoreQueryParams().setStorageQuotaInByte(Store.UNLIMITED_STORAGE_QUOTA).setChunkingEnabled(true);
-        updateStoreResponse =
-            parentControllerClient.retryableRequest(5, c -> c.updateStore(storeName, newUpdateStoreParams));
-        assertFalse(updateStoreResponse.isError(), "Update store got error: " + updateStoreResponse.getError());
-
-        // Perform one time repush to make sure repush can handle chunks data correctly.
-        // intentionally stop re-consuming from RT so stale records don't affect the testing results
-        props.put(REWIND_TIME_IN_SECONDS_OVERRIDE, 0);
-        IntegrationTestPushUtils.runVPJ(props);
-
-        TestUtils.waitForNonDeterministicPushCompletion(
-            Version.composeKafkaTopic(storeName, 3),
-            parentControllerClient,
-            30,
-            TimeUnit.SECONDS);
-
-        TestUtils.waitForNonDeterministicAssertion(10, TimeUnit.SECONDS, true, () -> {
-          try {
-            GenericRecord value = (GenericRecord) storeReader.get(keyRecord).get();
-            assertNotNull(value, "key " + keyRecord + " should not be missing!");
-          } catch (Exception e) {
-            throw new VeniceException(e);
-          }
-        });
-        partialUpdateRecord = new UpdateBuilderImpl(writeComputeSchema)
-            .setElementsToAddToListField("blockedContentsUrns", Collections.singletonList("urn:li:lyndaCourse:1"))
-            .build();
-        sendStreamingRecord(veniceProducer, storeName, keyRecord, partialUpdateRecord);
-        partialUpdateRecord = new UpdateBuilderImpl(writeComputeSchema)
-            .setElementsToAddToListField("blockedContentsUrns", Collections.singletonList("urn:li:lyndaCourse:2"))
-            .build();
-        sendStreamingRecord(veniceProducer, storeName, keyRecord, partialUpdateRecord);
-        sendStreamingRecord(veniceProducer, storeName, checkpointKeyRecord, partialUpdateRecord);
-        TestUtils.waitForNonDeterministicAssertion(10, TimeUnit.SECONDS, true, () -> {
-          try {
-            GenericRecord value = (GenericRecord) storeReader.get(checkpointKeyRecord).get();
-            assertNotNull(value, "key " + checkpointKeyRecord + " should not be missing!");
-          } catch (Exception e) {
-            throw new VeniceException(e);
-          }
-        });
-        TestUtils.waitForNonDeterministicAssertion(10, TimeUnit.SECONDS, true, () -> {
-          try {
-            GenericRecord value = (GenericRecord) storeReader.get(keyRecord).get();
-            assertNotNull(value, "key " + keyRecord + " should not be missing!");
-          } catch (Exception e) {
-            throw new VeniceException(e);
-          }
-        });
-      }
-    }
-  }
-
-  @Test(timeOut = TEST_TIMEOUT_MS)
-  public void testIncrementalPushPartialUpdateClassicFormat() throws IOException {
-    final String storeName = Utils.getUniqueString("inc_push_update_classic_format");
-    String parentControllerUrl = parentController.getControllerUrl();
-    File inputDir = getTempDataDirectory();
-    Schema recordSchema = writeSimpleAvroFileWithStringToPartialUpdateOpRecordSchema(inputDir);
-    String keySchemaStr = recordSchema.getField(DEFAULT_KEY_FIELD_PROP).schema().toString();
-    String inputDirPath = "file://" + inputDir.getAbsolutePath();
-    Properties vpjProperties =
-        IntegrationTestPushUtils.defaultVPJProps(multiRegionMultiClusterWrapper, inputDirPath, storeName);
-    vpjProperties.put(ENABLE_WRITE_COMPUTE, true);
-    vpjProperties.put(INCREMENTAL_PUSH, true);
-
-    try (ControllerClient parentControllerClient = new ControllerClient(CLUSTER_NAME, parentControllerUrl)) {
-      assertCommand(
-          parentControllerClient
-              .createNewStore(storeName, "test_owner", keySchemaStr, TestWriteUtils.NAME_RECORD_V1_SCHEMA.toString()));
-      UpdateStoreQueryParams updateStoreParams =
-          new UpdateStoreQueryParams().setStorageQuotaInByte(Store.UNLIMITED_STORAGE_QUOTA)
-              .setCompressionStrategy(CompressionStrategy.NO_OP)
-              .setWriteComputationEnabled(true)
-              .setChunkingEnabled(true)
-              .setIncrementalPushEnabled(true)
-              .setActiveActiveReplicationEnabled(true)
-              .setSeparateRealTimeTopicEnabled(true)
-              .setHybridRewindSeconds(10L)
-              .setHybridOffsetLagThreshold(2L);
-      ControllerResponse updateStoreResponse =
-          parentControllerClient.retryableRequest(5, c -> c.updateStore(storeName, updateStoreParams));
-      assertFalse(updateStoreResponse.isError(), "Update store got error: " + updateStoreResponse.getError());
-
-      VersionCreationResponse response = parentControllerClient.emptyPush(storeName, "test_push_id", 1000);
-      assertEquals(response.getVersion(), 1);
-      assertFalse(response.isError(), "Empty push to parent colo should succeed");
-      TestUtils.waitForNonDeterministicPushCompletion(
-          Version.composeKafkaTopic(storeName, 1),
-          parentControllerClient,
-          60,
-          TimeUnit.SECONDS);
-
-      // VPJ push
-      String childControllerUrl = childDatacenters.get(0).getRandomController().getControllerUrl();
-      try (ControllerClient childControllerClient = new ControllerClient(CLUSTER_NAME, childControllerUrl)) {
-        IntegrationTestPushUtils.runVPJ(vpjProperties, 1, childControllerClient);
-      }
-      VeniceClusterWrapper veniceClusterWrapper = childDatacenters.get(0).getClusters().get(CLUSTER_NAME);
-      veniceClusterWrapper.waitVersion(storeName, 1);
-      try (AvroGenericStoreClient<Object, Object> storeReader = ClientFactory.getAndStartGenericAvroClient(
-          ClientConfig.defaultGenericClientConfig(storeName).setVeniceURL(veniceClusterWrapper.getRandomRouterURL()))) {
-        TestUtils.waitForNonDeterministicAssertion(10, TimeUnit.SECONDS, true, () -> {
-          try {
-            for (int i = 1; i < 100; i++) {
-              String key = String.valueOf(i);
-              GenericRecord value = readValue(storeReader, key);
-              assertNotNull(value, "Key " + key + " should not be missing!");
-              assertEquals(value.get("firstName").toString(), "first_name_" + key);
-              assertEquals(value.get("lastName").toString(), "last_name_" + key);
-            }
-          } catch (Exception e) {
-            throw new VeniceException(e);
-          }
-        });
-      }
-    }
   }
 
   @Test(timeOut = TEST_TIMEOUT_MS)
@@ -466,18 +236,20 @@ public class PartialUpdateTest {
           TimeUnit.SECONDS);
     }
     VeniceClusterWrapper veniceClusterWrapper = childDatacenters.get(0).getClusters().get(CLUSTER_NAME);
-    SystemProducer veniceProducer = getSamzaProducer(veniceClusterWrapper, storeName, Version.PushType.STREAM);
     String key = "highway";
-    // Send first value with old schema;
-    GenericRecord record = new GenericData.Record(schemaV1);
-    record.put("id", "101");
-    record.put("name", "U.S. 101");
-    sendStreamingRecord(veniceProducer, storeName, key, record);
-    // Send second value with new schema; Without default schema carry fix, it will fail SIT ingestion.
-    GenericRecord recordV2 = new GenericData.Record(schemaV2);
-    recordV2.put("id", "280");
-    recordV2.put("name", "Interstate 280");
-    sendStreamingRecord(veniceProducer, storeName, key, recordV2);
+    try (VeniceSystemProducer veniceProducer =
+        getSamzaProducer(veniceClusterWrapper, storeName, Version.PushType.STREAM)) {
+      // Send first value with old schema;
+      GenericRecord record = new GenericData.Record(schemaV1);
+      record.put("id", "101");
+      record.put("name", "U.S. 101");
+      sendStreamingRecord(veniceProducer, storeName, key, record);
+      // Send second value with new schema; Without default schema carry fix, it will fail SIT ingestion.
+      GenericRecord recordV2 = new GenericData.Record(schemaV2);
+      recordV2.put("id", "280");
+      recordV2.put("name", "Interstate 280");
+      sendStreamingRecord(veniceProducer, storeName, key, recordV2);
+    }
 
     try (AvroGenericStoreClient<Object, Object> storeReader = ClientFactory.getAndStartGenericAvroClient(
         ClientConfig.defaultGenericClientConfig(storeName).setVeniceURL(veniceClusterWrapper.getRandomRouterURL()))) {
@@ -491,77 +263,6 @@ public class PartialUpdateTest {
           throw new VeniceException(e);
         }
       });
-    }
-  }
-
-  @Test(timeOut = TEST_TIMEOUT_MS, dataProvider = "True-and-False", dataProviderClass = DataProviderUtils.class)
-  public void testIncrementalPushPartialUpdateNewFormat(boolean useSparkCompute) throws IOException {
-    final String storeName = Utils.getUniqueString("inc_push_update_new_format");
-    String parentControllerUrl = parentController.getControllerUrl();
-    File inputDir = getTempDataDirectory();
-    Schema recordSchema = writeSimpleAvroFileWithStringToNameRecordV1Schema(inputDir);
-    String keySchemaStr = recordSchema.getField(DEFAULT_KEY_FIELD_PROP).schema().toString();
-    String inputDirPath = "file://" + inputDir.getAbsolutePath();
-    Properties vpjProperties =
-        IntegrationTestPushUtils.defaultVPJProps(multiRegionMultiClusterWrapper, inputDirPath, storeName);
-    vpjProperties.put(ENABLE_WRITE_COMPUTE, true);
-    vpjProperties.put(INCREMENTAL_PUSH, true);
-    if (useSparkCompute) {
-      vpjProperties.setProperty(DATA_WRITER_COMPUTE_JOB_CLASS, DataWriterSparkJob.class.getCanonicalName());
-      vpjProperties.setProperty(SPARK_NATIVE_INPUT_FORMAT_ENABLED, String.valueOf(true));
-    }
-
-    try (ControllerClient parentControllerClient = new ControllerClient(CLUSTER_NAME, parentControllerUrl)) {
-      assertCommand(
-          parentControllerClient
-              .createNewStore(storeName, "test_owner", keySchemaStr, NAME_RECORD_V2_SCHEMA.toString()));
-      UpdateStoreQueryParams updateStoreParams =
-          new UpdateStoreQueryParams().setStorageQuotaInByte(Store.UNLIMITED_STORAGE_QUOTA)
-              .setCompressionStrategy(CompressionStrategy.NO_OP)
-              .setActiveActiveReplicationEnabled(true)
-              .setWriteComputationEnabled(true)
-              .setChunkingEnabled(true)
-              .setIncrementalPushEnabled(true)
-              .setHybridRewindSeconds(10L)
-              .setHybridOffsetLagThreshold(2L);
-      ControllerResponse updateStoreResponse =
-          parentControllerClient.retryableRequest(5, c -> c.updateStore(storeName, updateStoreParams));
-      assertFalse(updateStoreResponse.isError(), "Update store got error: " + updateStoreResponse.getError());
-
-      VersionCreationResponse response = parentControllerClient.emptyPush(storeName, "test_push_id", 1000);
-      assertEquals(response.getVersion(), 1);
-      assertFalse(response.isError(), "Empty push to parent colo should succeed");
-      TestUtils.waitForNonDeterministicPushCompletion(
-          response.getKafkaTopic(),
-          parentControllerClient,
-          30,
-          TimeUnit.SECONDS);
-
-      // VPJ push
-      String childControllerUrl = childDatacenters.get(0).getRandomController().getControllerUrl();
-      try (ControllerClient childControllerClient = new ControllerClient(CLUSTER_NAME, childControllerUrl)) {
-        IntegrationTestPushUtils.runVPJ(vpjProperties, 1, childControllerClient);
-      }
-      VeniceClusterWrapper veniceClusterWrapper = childDatacenters.get(0).getClusters().get(CLUSTER_NAME);
-      veniceClusterWrapper.waitVersion(storeName, 1);
-
-      try (AvroGenericStoreClient<Object, Object> storeReader = ClientFactory.getAndStartGenericAvroClient(
-          ClientConfig.defaultGenericClientConfig(storeName).setVeniceURL(veniceClusterWrapper.getRandomRouterURL()))) {
-        TestUtils.waitForNonDeterministicAssertion(10, TimeUnit.SECONDS, true, () -> {
-          try {
-            for (int i = 1; i < 100; i++) {
-              String key = String.valueOf(i);
-              GenericRecord value = readValue(storeReader, key);
-              assertNotNull(value, "Key " + key + " should not be missing!");
-              assertEquals(value.get("firstName").toString(), "first_name_" + key);
-              assertEquals(value.get("lastName").toString(), "last_name_" + key);
-              assertEquals(value.get("age"), -1);
-            }
-          } catch (Exception e) {
-            throw new VeniceException(e);
-          }
-        });
-      }
     }
   }
 
@@ -610,11 +311,13 @@ public class PartialUpdateTest {
       }
       veniceClusterWrapper.waitVersion(storeName, 1);
       // Produce partial updates on batch pushed keys
-      SystemProducer veniceProducer = getSamzaProducer(veniceClusterWrapper, storeName, Version.PushType.STREAM);
-      for (int i = 1; i < 100; i++) {
-        GenericRecord partialUpdateRecord =
-            new UpdateBuilderImpl(writeComputeSchema).setNewFieldValue("firstName", "new_name_" + i).build();
-        sendStreamingRecord(veniceProducer, storeName, String.valueOf(i), partialUpdateRecord);
+      try (VeniceSystemProducer veniceProducer =
+          getSamzaProducer(veniceClusterWrapper, storeName, Version.PushType.STREAM)) {
+        for (int i = 1; i < 100; i++) {
+          GenericRecord partialUpdateRecord =
+              new UpdateBuilderImpl(writeComputeSchema).setNewFieldValue("firstName", "new_name_" + i).build();
+          sendStreamingRecord(veniceProducer, storeName, String.valueOf(i), partialUpdateRecord);
+        }
       }
 
       try (AvroGenericStoreClient<Object, Object> storeReader = ClientFactory.getAndStartGenericAvroClient(
@@ -637,7 +340,7 @@ public class PartialUpdateTest {
   }
 
   @Test(timeOut = TEST_TIMEOUT_MS)
-  public void testPartialUpdateOnBatchPushedChunkKeys() throws IOException {
+  public void testActiveActivePartialUpdateOnBatchPushedChunkKeys() throws IOException {
     final String storeName = Utils.getUniqueString("updateBatch");
     String parentControllerUrl = parentController.getControllerUrl();
     File inputDir = getTempDataDirectory();
@@ -676,11 +379,13 @@ public class PartialUpdateTest {
       }
       veniceClusterWrapper.waitVersion(storeName, 1);
       // Produce partial updates on batch pushed keys
-      SystemProducer veniceProducer = getSamzaProducer(veniceClusterWrapper, storeName, Version.PushType.STREAM);
-      for (int i = 1; i < 100; i++) {
-        GenericRecord partialUpdateRecord =
-            new UpdateBuilderImpl(writeComputeSchema).setNewFieldValue("key", "new_name_" + i).build();
-        sendStreamingRecord(veniceProducer, storeName, String.valueOf(i), partialUpdateRecord);
+      try (VeniceSystemProducer veniceProducer =
+          getSamzaProducer(veniceClusterWrapper, storeName, Version.PushType.STREAM)) {
+        for (int i = 1; i < 100; i++) {
+          GenericRecord partialUpdateRecord =
+              new UpdateBuilderImpl(writeComputeSchema).setNewFieldValue("key", "new_name_" + i).build();
+          sendStreamingRecord(veniceProducer, storeName, String.valueOf(i), partialUpdateRecord);
+        }
       }
 
       try (AvroGenericStoreClient<Object, Object> storeReader = ClientFactory.getAndStartGenericAvroClient(
@@ -741,7 +446,6 @@ public class PartialUpdateTest {
     }
 
     VeniceClusterWrapper veniceCluster = childDatacenters.get(0).getClusters().get(CLUSTER_NAME);
-    SystemProducer veniceProducer = getSamzaProducer(veniceCluster, storeName, Version.PushType.STREAM);
 
     String key = "key1";
     String primitiveFieldName = "name";
@@ -750,8 +454,9 @@ public class PartialUpdateTest {
     // Insert large amount of Map entries to trigger RMD chunking.
     int oldUpdateCount = 29;
     int singleUpdateEntryCount = 10000;
-    try (AvroGenericStoreClient<Object, Object> storeReader = ClientFactory.getAndStartGenericAvroClient(
-        ClientConfig.defaultGenericClientConfig(storeName).setVeniceURL(veniceCluster.getRandomRouterURL()))) {
+    try (VeniceSystemProducer veniceProducer = getSamzaProducer(veniceCluster, storeName, Version.PushType.STREAM);
+        AvroGenericStoreClient<Object, Object> storeReader = ClientFactory.getAndStartGenericAvroClient(
+            ClientConfig.defaultGenericClientConfig(storeName).setVeniceURL(veniceCluster.getRandomRouterURL()))) {
       for (int i = 0; i < oldUpdateCount; i++) {
         producePartialUpdate(
             storeName,
@@ -823,8 +528,6 @@ public class PartialUpdateTest {
           Assert.assertNull(valueChunkBytes);
         }, false);
       });
-    } finally {
-      veniceProducer.stop();
     }
   }
 
@@ -882,7 +585,6 @@ public class PartialUpdateTest {
     }
 
     VeniceClusterWrapper veniceCluster = childDatacenters.get(0).getClusters().get(CLUSTER_NAME);
-    SystemProducer veniceProducer = getSamzaProducer(veniceCluster, storeName, Version.PushType.STREAM);
 
     String key = "key1";
     String primitiveFieldName = "name";
@@ -891,8 +593,9 @@ public class PartialUpdateTest {
     int totalUpdateCount = 40;
     // Insert large amount of Map entries to trigger RMD chunking.
     int singleUpdateEntryCount = 10000;
-    try (AvroGenericStoreClient<Object, Object> storeReader = ClientFactory.getAndStartGenericAvroClient(
-        ClientConfig.defaultGenericClientConfig(storeName).setVeniceURL(veniceCluster.getRandomRouterURL()))) {
+    try (VeniceSystemProducer veniceProducer = getSamzaProducer(veniceCluster, storeName, Version.PushType.STREAM);
+        AvroGenericStoreClient<Object, Object> storeReader = ClientFactory.getAndStartGenericAvroClient(
+            ClientConfig.defaultGenericClientConfig(storeName).setVeniceURL(veniceCluster.getRandomRouterURL()))) {
       for (int i = 0; i < (totalUpdateCount - 1); i++) {
         producePartialUpdateToArray(
             storeName,
@@ -1053,8 +756,6 @@ public class PartialUpdateTest {
         GenericRecord collectionFieldTimestampRecord = (GenericRecord) timestampRecord.get(listFieldName);
         assertEquals(collectionFieldTimestampRecord.get(TOP_LEVEL_TS_FIELD_NAME), (long) (totalUpdateCount) * 10);
       });
-    } finally {
-      veniceProducer.stop();
     }
 
     String metricName = AbstractVeniceStats.getSensorFullName(storeName, ASSEMBLED_RMD_SIZE_IN_BYTES) + ".Max";
@@ -1063,7 +764,7 @@ public class PartialUpdateTest {
   }
 
   @Test(timeOut = TEST_TIMEOUT_MS * 3)
-  public void testNonAARewind() throws Exception {
+  public void testNonAARewind() {
     final String storeName = Utils.getUniqueString("test");
     String parentControllerUrl = parentController.getControllerUrl();
     String keySchemaStr = "{\"type\" : \"string\"}";
@@ -1093,7 +794,6 @@ public class PartialUpdateTest {
     }
 
     VeniceClusterWrapper veniceCluster = childDatacenters.get(0).getClusters().get(CLUSTER_NAME);
-    SystemProducer veniceProducer = getSamzaProducer(veniceCluster, storeName, Version.PushType.STREAM);
 
     String key = "key";
     String primitiveFieldName = "name";
@@ -1101,8 +801,9 @@ public class PartialUpdateTest {
     String mapFieldName = "stringMap";
 
     int totalUpdateCount = 1000;
-    try (AvroGenericStoreClient<Object, Object> storeReader = ClientFactory.getAndStartGenericAvroClient(
-        ClientConfig.defaultGenericClientConfig(storeName).setVeniceURL(veniceCluster.getRandomRouterURL()))) {
+    try (VeniceSystemProducer veniceProducer = getSamzaProducer(veniceCluster, storeName, Version.PushType.STREAM);
+        AvroGenericStoreClient<Object, Object> storeReader = ClientFactory.getAndStartGenericAvroClient(
+            ClientConfig.defaultGenericClientConfig(storeName).setVeniceURL(veniceCluster.getRandomRouterURL()))) {
       for (int i = 0; i < totalUpdateCount; i++) {
         GenericRecord value = new GenericData.Record(valueSchema);
         value.put(primitiveFieldName, "London");
@@ -1123,10 +824,8 @@ public class PartialUpdateTest {
           throw new VeniceException(e);
         }
       });
-
-    } finally {
-      veniceProducer.stop();
     }
+
     try (ControllerClient parentControllerClient = new ControllerClient(CLUSTER_NAME, parentControllerUrl)) {
       VersionCreationResponse response =
           TestUtils.assertCommand(parentControllerClient.emptyPush(storeName, "test_push_id_v2", 1000));
@@ -1146,240 +845,9 @@ public class PartialUpdateTest {
     Assert.assertTrue(totalCountMetric <= 2.0d);
   }
 
-  @Test(timeOut = TEST_TIMEOUT_MS, dataProvider = "Compression-Strategies", dataProviderClass = DataProviderUtils.class)
-  public void testRepushWithTTLWithActiveActivePartialUpdateStore(CompressionStrategy compressionStrategy) {
-    final String storeName = Utils.getUniqueString("ttlRepushAAWC");
-    String parentControllerUrl = parentController.getControllerUrl();
-    Schema valueSchema = AvroCompatibilityHelper.parse(loadFileAsString("CollectionRecordV1.avsc"));
-    Schema partialUpdateSchema = WriteComputeSchemaConverter.getInstance().convertFromValueRecordSchema(valueSchema);
-
-    Schema rmdSchema = RmdSchemaGenerator.generateMetadataSchema(valueSchema);
-    ReadOnlySchemaRepository schemaRepo = mock(ReadOnlySchemaRepository.class);
-    when(schemaRepo.getReplicationMetadataSchema(storeName, 1, 1)).thenReturn(new RmdSchemaEntry(1, 1, rmdSchema));
-    when(schemaRepo.getDerivedSchema(storeName, 1, 1)).thenReturn(new DerivedSchemaEntry(1, 1, partialUpdateSchema));
-    when(schemaRepo.getValueSchema(storeName, 1)).thenReturn(new SchemaEntry(1, valueSchema));
-    StringAnnotatedStoreSchemaCache stringAnnotatedStoreSchemaCache =
-        new StringAnnotatedStoreSchemaCache(storeName, schemaRepo);
-    RmdSerDe rmdSerDe = new RmdSerDe(stringAnnotatedStoreSchemaCache, 1);
-
-    try (ControllerClient parentControllerClient = new ControllerClient(CLUSTER_NAME, parentControllerUrl)) {
-      assertCommand(
-          parentControllerClient
-              .createNewStore(storeName, "test_owner", STRING_SCHEMA.toString(), valueSchema.toString()));
-      UpdateStoreQueryParams updateStoreParams =
-          new UpdateStoreQueryParams().setStorageQuotaInByte(Store.UNLIMITED_STORAGE_QUOTA)
-              .setPartitionCount(1)
-              .setCompressionStrategy(compressionStrategy)
-              .setWriteComputationEnabled(true)
-              .setActiveActiveReplicationEnabled(true)
-              .setChunkingEnabled(true)
-              .setRmdChunkingEnabled(true)
-              .setHybridRewindSeconds(1L)
-              .setHybridOffsetLagThreshold(1L);
-      ControllerResponse updateStoreResponse =
-          parentControllerClient.retryableRequest(5, c -> c.updateStore(storeName, updateStoreParams));
-      assertFalse(updateStoreResponse.isError(), "Update store got error: " + updateStoreResponse.getError());
-
-      VersionCreationResponse response = parentControllerClient.emptyPush(storeName, "test_push_id", 1000);
-      assertEquals(response.getVersion(), 1);
-      assertFalse(response.isError(), "Empty push to parent colo should succeed");
-      TestUtils.waitForNonDeterministicPushCompletion(
-          Version.composeKafkaTopic(storeName, 1),
-          parentControllerClient,
-          30,
-          TimeUnit.SECONDS);
-    }
-
-    VeniceClusterWrapper veniceCluster = childDatacenters.get(0).getClusters().get(CLUSTER_NAME);
-    SystemProducer veniceProducer = getSamzaProducer(veniceCluster, storeName, Version.PushType.STREAM);
-    long STALE_TS = 99999L;
-    long FRESH_TS = 100000L;
-    String STRING_MAP_FILED = "stringMap";
-    String REGULAR_FIELD = "name";
-    /**
-     * Case 1: The record is partially stale, TTL repush should only keep the part that's fresh.
-     */
-    String key1 = "key1";
-    // This update is expected to be carried into TTL repush.
-    UpdateBuilder updateBuilder = new UpdateBuilderImpl(partialUpdateSchema);
-    updateBuilder.setEntriesToAddToMapField(STRING_MAP_FILED, Collections.singletonMap("k1", "v1"));
-    sendStreamingRecord(veniceProducer, storeName, key1, updateBuilder.build(), FRESH_TS);
-    // This update is expected to be WIPED OUT after TTL repush.
-    updateBuilder = new UpdateBuilderImpl(partialUpdateSchema);
-    updateBuilder.setNewFieldValue(REGULAR_FIELD, "new_name");
-    updateBuilder.setEntriesToAddToMapField(STRING_MAP_FILED, Collections.singletonMap("k2", "v2"));
-    sendStreamingRecord(veniceProducer, storeName, key1, updateBuilder.build(), STALE_TS);
-
-    /**
-     * Case 2: The record is fully stale, TTL repush should drop the record.
-     */
-    String key2 = "key2";
-    // This update is expected to be WIPED OUT after TTL repush.
-    updateBuilder = new UpdateBuilderImpl(partialUpdateSchema);
-    updateBuilder.setNewFieldValue(REGULAR_FIELD, "new_name_2");
-    sendStreamingRecord(veniceProducer, storeName, key2, updateBuilder.build(), STALE_TS);
-
-    /**
-     * Case 3: The record is fully fresh, TTL repush should keep the record.
-     */
-    String key3 = "key3";
-    // This update is expected to be carried into TTL repush.
-    updateBuilder = new UpdateBuilderImpl(partialUpdateSchema);
-    updateBuilder.setNewFieldValue(REGULAR_FIELD, "new_name_3");
-    sendStreamingRecord(veniceProducer, storeName, key3, updateBuilder.build(), FRESH_TS);
-
-    /**
-     * Case 4: The delete record is fully fresh, TTL repush should NOT FAIL due to this record.
-     */
-    String key4 = "key4";
-    sendStreamingRecord(veniceProducer, storeName, key4, updateBuilder.build(), FRESH_TS);
-    sendStreamingDeleteRecord(veniceProducer, storeName, key4, FRESH_TS + 1);
-
-    /**
-     * Case 5: The delete record is stale, TTL repush should keep new update.
-     */
-    String key5 = "key5";
-    sendStreamingDeleteRecord(veniceProducer, storeName, key5, STALE_TS);
-    updateBuilder = new UpdateBuilderImpl(partialUpdateSchema);
-    updateBuilder.setNewFieldValue(REGULAR_FIELD, "new_name_5");
-    sendStreamingRecord(veniceProducer, storeName, key5, updateBuilder.build(), FRESH_TS);
-
-    /**
-     * Validate the data is ready in storage before TTL repush.
-     */
-    try (AvroGenericStoreClient<Object, Object> storeReader = ClientFactory.getAndStartGenericAvroClient(
-        ClientConfig.defaultGenericClientConfig(storeName).setVeniceURL(veniceCluster.getRandomRouterURL()))) {
-      TestUtils.waitForNonDeterministicAssertion(ASSERTION_TIMEOUT_MS, TimeUnit.MILLISECONDS, true, () -> {
-        try {
-          GenericRecord valueRecord = readValue(storeReader, key1);
-          assertNotNull(valueRecord);
-          assertEquals(valueRecord.get(REGULAR_FIELD), new Utf8("new_name"));
-          assertNotNull(valueRecord.get(STRING_MAP_FILED));
-          Map<Utf8, Utf8> stringMapValue = (Map<Utf8, Utf8>) valueRecord.get(STRING_MAP_FILED);
-          assertEquals(stringMapValue.get(new Utf8("k1")), new Utf8("v1"));
-          assertEquals(stringMapValue.get(new Utf8("k2")), new Utf8("v2"));
-
-          valueRecord = readValue(storeReader, key2);
-          assertNotNull(valueRecord);
-          assertEquals(valueRecord.get(REGULAR_FIELD), new Utf8("new_name_2"));
-
-          valueRecord = readValue(storeReader, key3);
-          assertNotNull(valueRecord);
-          assertEquals(valueRecord.get(REGULAR_FIELD), new Utf8("new_name_3"));
-
-          valueRecord = readValue(storeReader, key4);
-          assertNull(valueRecord);
-
-          valueRecord = readValue(storeReader, key5);
-          assertNotNull(valueRecord);
-          assertEquals(valueRecord.get(REGULAR_FIELD), new Utf8("new_name_5"));
-        } catch (Exception e) {
-          throw new VeniceException(e);
-        }
-      });
-    }
-
-    String kafkaTopic_v1 = Version.composeKafkaTopic(storeName, 1);
-    // Validate RMD bytes after initial update requests.
-    validateRmdData(rmdSerDe, kafkaTopic_v1, key1, rmdWithValueSchemaId -> {
-      GenericRecord timestampRecord = (GenericRecord) rmdWithValueSchemaId.getRmdRecord().get(TIMESTAMP_FIELD_NAME);
-      GenericRecord stringMapTsRecord = (GenericRecord) timestampRecord.get(STRING_MAP_FILED);
-      Assert.assertEquals(stringMapTsRecord.get(TOP_LEVEL_TS_FIELD_NAME), 0L);
-      Assert.assertEquals(stringMapTsRecord.get(ACTIVE_ELEM_TS_FIELD_NAME), Arrays.asList(STALE_TS, FRESH_TS));
-      Assert.assertEquals(timestampRecord.get(REGULAR_FIELD), STALE_TS);
-    });
-
-    validateRmdData(rmdSerDe, kafkaTopic_v1, key2, rmdWithValueSchemaId -> {
-      GenericRecord timestampRecord = (GenericRecord) rmdWithValueSchemaId.getRmdRecord().get(TIMESTAMP_FIELD_NAME);
-      Assert.assertEquals(timestampRecord.get(REGULAR_FIELD), STALE_TS);
-    });
-
-    validateRmdData(rmdSerDe, kafkaTopic_v1, key3, rmdWithValueSchemaId -> {
-      GenericRecord timestampRecord = (GenericRecord) rmdWithValueSchemaId.getRmdRecord().get(TIMESTAMP_FIELD_NAME);
-      Assert.assertEquals(timestampRecord.get(REGULAR_FIELD), FRESH_TS);
-    });
-
-    // Perform one time repush to make sure repush can handle RMD chunks data correctly.
-    Properties props =
-        IntegrationTestPushUtils.defaultVPJProps(multiRegionMultiClusterWrapper, "dummyInputPath", storeName);
-    props.setProperty(SOURCE_KAFKA, "true");
-    props.setProperty(KAFKA_INPUT_BROKER_URL, veniceCluster.getPubSubBrokerWrapper().getAddress());
-    props.setProperty(KAFKA_INPUT_MAX_RECORDS_PER_MAPPER, "5");
-    props.setProperty(REPUSH_TTL_ENABLE, "true");
-    // Override the TTL repush start TS to work with logical TS setup.
-    props.setProperty(REPUSH_TTL_START_TIMESTAMP, String.valueOf(FRESH_TS));
-    // Override the rewind time to make sure not to consume 24hrs data from RT topic.
-    props.put(REWIND_TIME_IN_SECONDS_OVERRIDE, 0);
-    IntegrationTestPushUtils.runVPJ(props);
-    try (ControllerClient parentControllerClient = new ControllerClient(CLUSTER_NAME, parentControllerUrl)) {
-      TestUtils.waitForNonDeterministicPushCompletion(
-          Version.composeKafkaTopic(storeName, 2),
-          parentControllerClient,
-          30,
-          TimeUnit.SECONDS);
-    }
-
-    /**
-     * Validate the data is ready in storage after TTL repush.
-     */
-    try (AvroGenericStoreClient<Object, Object> storeReader = ClientFactory.getAndStartGenericAvroClient(
-        ClientConfig.defaultGenericClientConfig(storeName).setVeniceURL(veniceCluster.getRandomRouterURL()))) {
-      TestUtils.waitForNonDeterministicAssertion(ASSERTION_TIMEOUT_MS, TimeUnit.MILLISECONDS, true, () -> {
-        try {
-          // Key 1 is partially preserved.
-          GenericRecord valueRecord = readValue(storeReader, key1);
-          assertNotNull(valueRecord);
-          assertEquals(valueRecord.get(REGULAR_FIELD), new Utf8("default_name"));
-          assertNotNull(valueRecord.get(STRING_MAP_FILED));
-          Map<Utf8, Utf8> stringMapValue = (Map<Utf8, Utf8>) valueRecord.get(STRING_MAP_FILED);
-          assertEquals(stringMapValue.get(new Utf8("k1")), new Utf8("v1"));
-          assertNull(stringMapValue.get(new Utf8("k2")));
-
-          // Key 2 is fully removed.
-          valueRecord = readValue(storeReader, key2);
-          assertNull(valueRecord);
-
-          // Key 3 is fully preserved.
-          valueRecord = readValue(storeReader, key3);
-          assertNotNull(valueRecord);
-          assertEquals(valueRecord.get(REGULAR_FIELD), new Utf8("new_name_3"));
-
-          valueRecord = readValue(storeReader, key4);
-          assertNull(valueRecord);
-
-          valueRecord = readValue(storeReader, key5);
-          assertNotNull(valueRecord);
-          assertEquals(valueRecord.get(REGULAR_FIELD), new Utf8("new_name_5"));
-        } catch (Exception e) {
-          throw new VeniceException(e);
-        }
-      });
-    }
-
-    String kafkaTopic_v2 = Version.composeKafkaTopic(storeName, 2);
-    // Validate RMD bytes after TTL repush.
-    validateRmdData(rmdSerDe, kafkaTopic_v2, key1, rmdWithValueSchemaId -> {
-      GenericRecord timestampRecord = (GenericRecord) rmdWithValueSchemaId.getRmdRecord().get(TIMESTAMP_FIELD_NAME);
-      GenericRecord stringMapTsRecord = (GenericRecord) timestampRecord.get(STRING_MAP_FILED);
-      Assert.assertEquals(stringMapTsRecord.get(TOP_LEVEL_TS_FIELD_NAME), STALE_TS);
-      Assert.assertEquals(stringMapTsRecord.get(ACTIVE_ELEM_TS_FIELD_NAME), Collections.singletonList(FRESH_TS));
-      Assert.assertEquals(timestampRecord.get(REGULAR_FIELD), STALE_TS);
-    });
-    validateRmdData(rmdSerDe, kafkaTopic_v1, key2, rmdWithValueSchemaId -> {
-      GenericRecord timestampRecord = (GenericRecord) rmdWithValueSchemaId.getRmdRecord().get(TIMESTAMP_FIELD_NAME);
-      Assert.assertEquals(timestampRecord.get(REGULAR_FIELD), STALE_TS);
-    });
-
-    validateRmdData(rmdSerDe, kafkaTopic_v1, key3, rmdWithValueSchemaId -> {
-      GenericRecord timestampRecord = (GenericRecord) rmdWithValueSchemaId.getRmdRecord().get(TIMESTAMP_FIELD_NAME);
-      Assert.assertEquals(timestampRecord.get(REGULAR_FIELD), FRESH_TS);
-    });
-
-  }
-
   @Test(timeOut = TEST_TIMEOUT_MS)
   public void testEnablePartialUpdateOnActiveActiveStore() {
-    final String storeName = Utils.getUniqueString("testRepushWithDeleteRecord");
+    final String storeName = Utils.getUniqueString("testEnablePartialUpdateOnActiveActiveStore");
     String parentControllerUrl = parentController.getControllerUrl();
     Schema valueSchema = AvroCompatibilityHelper.parse(loadFileAsString("CollectionRecordV1.avsc"));
     Schema partialUpdateSchema = WriteComputeSchemaConverter.getInstance().convertFromValueRecordSchema(valueSchema);
@@ -1554,7 +1022,7 @@ public class PartialUpdateTest {
     }
 
     VeniceClusterWrapper veniceCluster = childDatacenters.get(0).getClusters().get(CLUSTER_NAME);
-    SystemProducer veniceProducer = getSamzaProducer(veniceCluster, storeName, Version.PushType.STREAM);
+
     String key1 = "key1";
 
     GenericRecord record = new GenericData.Record(valueSchema);
@@ -1562,7 +1030,9 @@ public class PartialUpdateTest {
     record.put("floatArray", Collections.emptyList());
     record.put("stringMap", Collections.emptyMap());
 
-    sendStreamingRecord(veniceProducer, storeName, key1, record, 10000L);
+    try (VeniceSystemProducer veniceProducer = getSamzaProducer(veniceCluster, storeName, Version.PushType.STREAM)) {
+      sendStreamingRecord(veniceProducer, storeName, key1, record, 10000L);
+    }
 
     for (int i = 0; i < NUMBER_OF_CHILD_DATACENTERS; i++) {
       VeniceClusterWrapper cluster = childDatacenters.get(i).getClusters().get(CLUSTER_NAME);
@@ -1645,10 +1115,12 @@ public class PartialUpdateTest {
           REPLICATION_FACTOR - 1);
     });
     // Need to create a new producer to update partial update config from store response.
-    veniceProducer = getSamzaProducer(veniceCluster, storeName, Version.PushType.STREAM);
     UpdateBuilder builder = new UpdateBuilderImpl(partialUpdateSchema);
     builder.setNewFieldValue("name", "field-level TS");
-    sendStreamingRecord(veniceProducer, storeName, key1, builder.build(), 10001L);
+    try (VeniceSystemProducer veniceProducer = getSamzaProducer(veniceCluster, storeName, Version.PushType.STREAM)) {
+      sendStreamingRecord(veniceProducer, storeName, key1, builder.build(), 10001L);
+    }
+
     for (int i = 0; i < NUMBER_OF_CHILD_DATACENTERS; i++) {
       VeniceClusterWrapper cluster = childDatacenters.get(i).getClusters().get(CLUSTER_NAME);
       try (AvroGenericStoreClient<Object, Object> storeReader = ClientFactory.getAndStartGenericAvroClient(
@@ -1664,38 +1136,6 @@ public class PartialUpdateTest {
         });
       }
     }
-  }
-
-  @Test(timeOut = TEST_TIMEOUT)
-  public void testRTTopicDeletionWithHybridAndIncrementalVersions() {
-    String storeName = Utils.getUniqueString("testRTTopicDeletion");
-    String parentControllerURLs = multiRegionMultiClusterWrapper.getControllerConnectString();
-    ControllerClient parentControllerClient =
-        ControllerClient.constructClusterControllerClient(CLUSTER_NAME, parentControllerURLs);
-
-    NewStoreResponse newStoreResponse =
-        parentControllerClient.retryableRequest(5, c -> c.createNewStore(storeName, "", "\"string\"", "\"string\""));
-    Assert.assertFalse(
-        newStoreResponse.isError(),
-        "The NewStoreResponse returned an error: " + newStoreResponse.getError());
-
-    UpdateStoreQueryParams updateStoreParams = new UpdateStoreQueryParams();
-    updateStoreParams.setBackupStrategy(BackupStrategy.KEEP_MIN_VERSIONS)
-        .setActiveActiveReplicationEnabled(true)
-        .setIncrementalPushEnabled(true)
-        .setNumVersionsToPreserve(2)
-        .setHybridRewindSeconds(1000)
-        .setHybridOffsetLagThreshold(1000);
-    TestWriteUtils.updateStore(storeName, parentControllerClient, updateStoreParams);
-
-    VersionCreationResponse response = parentControllerClient.emptyPush(storeName, "test_push_id", 1000);
-    assertEquals(response.getVersion(), 1);
-    assertFalse(response.isError(), "Empty push to parent colo should succeed");
-    TestUtils.waitForNonDeterministicPushCompletion(
-        Version.composeKafkaTopic(storeName, 1),
-        parentControllerClient,
-        60,
-        TimeUnit.SECONDS);
   }
 
   @Test(timeOut = TEST_TIMEOUT_MS)
@@ -1733,7 +1173,6 @@ public class PartialUpdateTest {
     }
 
     VeniceClusterWrapper veniceCluster = childDatacenters.get(0).getClusters().get(CLUSTER_NAME);
-    SystemProducer veniceProducer = getSamzaProducer(veniceCluster, storeName, Version.PushType.STREAM);
     String key1 = "key1";
     String key2 = "key2";
     String key3 = "key3";
@@ -1743,9 +1182,11 @@ public class PartialUpdateTest {
     record.put("floatArray", Collections.emptyList());
     record.put("stringMap", Collections.emptyMap());
 
-    sendStreamingRecord(veniceProducer, storeName, key1, record, 10000L);
-    sendStreamingRecord(veniceProducer, storeName, key2, record, 10000L);
-    sendStreamingRecord(veniceProducer, storeName, key3, record, 10000L);
+    try (VeniceSystemProducer veniceProducer = getSamzaProducer(veniceCluster, storeName, Version.PushType.STREAM)) {
+      sendStreamingRecord(veniceProducer, storeName, key1, record, 10000L);
+      sendStreamingRecord(veniceProducer, storeName, key2, record, 10000L);
+      sendStreamingRecord(veniceProducer, storeName, key3, record, 10000L);
+    }
 
     for (int i = 0; i < NUMBER_OF_CHILD_DATACENTERS; i++) {
       VeniceClusterWrapper cluster = childDatacenters.get(i).getClusters().get(CLUSTER_NAME);
@@ -1784,13 +1225,13 @@ public class PartialUpdateTest {
     record.put("floatArray", Collections.emptyList());
     record.put("stringMap", Collections.emptyMap());
     // New UPDATE record for KEY3
-    veniceProducer = getSamzaProducer(veniceCluster, storeName, Version.PushType.STREAM);
     UpdateBuilder builder = new UpdateBuilderImpl(partialUpdateSchema);
     builder.setNewFieldValue("name", "field-level TS UPDATE");
-
-    sendStreamingRecord(veniceProducer, storeName, key1, record, 10001L);
-    sendStreamingDeleteRecord(veniceProducer, storeName, key2, 10001L);
-    sendStreamingRecord(veniceProducer, storeName, key3, builder.build(), 10001L);
+    try (VeniceSystemProducer veniceProducer = getSamzaProducer(veniceCluster, storeName, Version.PushType.STREAM)) {
+      sendStreamingRecord(veniceProducer, storeName, key1, record, 10001L);
+      sendStreamingDeleteRecord(veniceProducer, storeName, key2, 10001L);
+      sendStreamingRecord(veniceProducer, storeName, key3, builder.build(), 10001L);
+    }
     for (int i = 0; i < NUMBER_OF_CHILD_DATACENTERS; i++) {
       VeniceClusterWrapper cluster = childDatacenters.get(i).getClusters().get(CLUSTER_NAME);
       try (AvroGenericStoreClient<Object, Object> storeReader = ClientFactory.getAndStartGenericAvroClient(
@@ -1812,183 +1253,6 @@ public class PartialUpdateTest {
           }
         });
       }
-    }
-  }
-
-  @Test(timeOut = TEST_TIMEOUT_MS)
-  public void testRepushWithDeleteRecord() {
-    final String storeName = Utils.getUniqueString("testRepushWithDeleteRecord");
-    String parentControllerUrl = parentController.getControllerUrl();
-    Schema valueSchema = AvroCompatibilityHelper.parse(loadFileAsString("CollectionRecordV1.avsc"));
-    Schema partialUpdateSchema = WriteComputeSchemaConverter.getInstance().convertFromValueRecordSchema(valueSchema);
-
-    try (ControllerClient parentControllerClient = new ControllerClient(CLUSTER_NAME, parentControllerUrl)) {
-      assertCommand(
-          parentControllerClient
-              .createNewStore(storeName, "test_owner", STRING_SCHEMA.toString(), valueSchema.toString()));
-      UpdateStoreQueryParams updateStoreParams =
-          new UpdateStoreQueryParams().setStorageQuotaInByte(Store.UNLIMITED_STORAGE_QUOTA)
-              .setPartitionCount(1)
-              .setCompressionStrategy(CompressionStrategy.NO_OP)
-              .setWriteComputationEnabled(true)
-              .setActiveActiveReplicationEnabled(true)
-              .setChunkingEnabled(true)
-              .setRmdChunkingEnabled(true)
-              .setHybridRewindSeconds(1L)
-              .setHybridOffsetLagThreshold(1L);
-      ControllerResponse updateStoreResponse =
-          parentControllerClient.retryableRequest(5, c -> c.updateStore(storeName, updateStoreParams));
-      assertFalse(updateStoreResponse.isError(), "Update store got error: " + updateStoreResponse.getError());
-
-      VersionCreationResponse response = parentControllerClient.emptyPush(storeName, "test_push_id", 1000);
-      assertEquals(response.getVersion(), 1);
-      assertFalse(response.isError(), "Empty push to parent colo should succeed");
-      TestUtils.waitForNonDeterministicPushCompletion(
-          Version.composeKafkaTopic(storeName, 1),
-          parentControllerClient,
-          30,
-          TimeUnit.SECONDS);
-    }
-
-    VeniceClusterWrapper veniceCluster = childDatacenters.get(0).getClusters().get(CLUSTER_NAME);
-    SystemProducer veniceProducer = getSamzaProducer(veniceCluster, storeName, Version.PushType.STREAM);
-    String STRING_MAP_FILED = "stringMap";
-    String key1 = "key1";
-    String key2 = "key2";
-    String key3 = "key3";
-
-    UpdateBuilder updateBuilder = new UpdateBuilderImpl(partialUpdateSchema);
-    updateBuilder.setEntriesToAddToMapField(STRING_MAP_FILED, Collections.singletonMap("k1", "v1"));
-    sendStreamingRecord(veniceProducer, storeName, key1, updateBuilder.build(), 10000L);
-    sendStreamingDeleteRecord(veniceProducer, storeName, key1, 10001L);
-    sendStreamingRecord(veniceProducer, storeName, key2, updateBuilder.build(), 10001L);
-
-    Properties props =
-        IntegrationTestPushUtils.defaultVPJProps(multiRegionMultiClusterWrapper, "dummyInputPath", storeName);
-    props.setProperty(SOURCE_KAFKA, "true");
-    props.setProperty(KAFKA_INPUT_BROKER_URL, veniceCluster.getPubSubBrokerWrapper().getAddress());
-    props.setProperty(KAFKA_INPUT_MAX_RECORDS_PER_MAPPER, "5");
-    IntegrationTestPushUtils.runVPJ(props);
-    try (ControllerClient parentControllerClient = new ControllerClient(CLUSTER_NAME, parentControllerUrl)) {
-      TestUtils.waitForNonDeterministicPushCompletion(
-          Version.composeKafkaTopic(storeName, 2),
-          parentControllerClient,
-          30,
-          TimeUnit.SECONDS);
-    }
-    for (int i = 0; i < NUMBER_OF_CHILD_DATACENTERS; i++) {
-      VeniceClusterWrapper cluster = childDatacenters.get(i).getClusters().get(CLUSTER_NAME);
-      try (AvroGenericStoreClient<Object, Object> storeReader = ClientFactory.getAndStartGenericAvroClient(
-          ClientConfig.defaultGenericClientConfig(storeName).setVeniceURL(cluster.getRandomRouterURL()))) {
-        TestUtils.waitForNonDeterministicAssertion(ASSERTION_TIMEOUT_MS, TimeUnit.MILLISECONDS, true, () -> {
-          try {
-            // Key 1 is deleted.
-            GenericRecord valueRecord = readValue(storeReader, key1);
-            assertNull(valueRecord);
-
-            // Key 2 is preserved.
-            valueRecord = readValue(storeReader, key2);
-            assertNotNull(valueRecord);
-            assertNotNull(valueRecord.get(STRING_MAP_FILED));
-            Map<Utf8, Utf8> stringMapValue = (Map<Utf8, Utf8>) valueRecord.get(STRING_MAP_FILED);
-            assertEquals(stringMapValue.get(new Utf8("k1")), new Utf8("v1"));
-          } catch (Exception e) {
-            throw new VeniceException(e);
-          }
-        });
-      }
-    }
-
-    String baseDataPath = Utils.getTempDataDirectory().getAbsolutePath();
-    VeniceProperties backendConfig = new PropertyBuilder().put(CLIENT_USE_SYSTEM_STORE_REPOSITORY, true)
-        .put(CLIENT_SYSTEM_STORE_REPOSITORY_REFRESH_INTERVAL_SECONDS, 1)
-        .put(DATA_BASE_PATH, baseDataPath)
-        .put(PERSISTENCE_TYPE, ROCKS_DB)
-        .put(ROCKSDB_PLAIN_TABLE_FORMAT_ENABLED, "false")
-        .put(ROCKSDB_BLOCK_CACHE_SIZE_IN_BYTES, 2 * 1024 * 1024L)
-        .put(DAVINCI_PUSH_STATUS_CHECK_INTERVAL_IN_MS, 1000)
-        .build();
-
-    MetricsRepository metricsRepository = new VeniceMetricsRepository();
-    try (CachingDaVinciClientFactory factory = new CachingDaVinciClientFactory(
-        d2ClientDC0,
-        VeniceRouterWrapper.CLUSTER_DISCOVERY_D2_SERVICE_NAME,
-        metricsRepository,
-        backendConfig)) {
-      DaVinciClient<Integer, Object> client =
-          factory.getAndStartGenericAvroClient(storeName, new DaVinciConfig().setStorageClass(StorageClass.DISK));
-      client.subscribeAll().get();
-    } catch (Exception e) {
-      throw new VeniceException(e);
-    }
-
-    sendStreamingRecord(veniceProducer, storeName, key1, updateBuilder.build(), 9999L);
-    sendStreamingRecord(veniceProducer, storeName, key3, updateBuilder.build(), 10000L);
-    for (int i = 0; i < NUMBER_OF_CHILD_DATACENTERS; i++) {
-      VeniceClusterWrapper cluster = childDatacenters.get(i).getClusters().get(CLUSTER_NAME);
-      try (AvroGenericStoreClient<Object, Object> storeReader = ClientFactory.getAndStartGenericAvroClient(
-          ClientConfig.defaultGenericClientConfig(storeName).setVeniceURL(cluster.getRandomRouterURL()))) {
-        TestUtils.waitForNonDeterministicAssertion(ASSERTION_TIMEOUT_MS, TimeUnit.MILLISECONDS, true, () -> {
-          try {
-            // Key 1 is deleted.
-            GenericRecord valueRecord = readValue(storeReader, key1);
-            assertNull(valueRecord);
-
-            // Key 3 is preserved.
-            valueRecord = readValue(storeReader, key3);
-            assertNotNull(valueRecord);
-            assertNotNull(valueRecord.get(STRING_MAP_FILED));
-            Map<Utf8, Utf8> stringMapValue = (Map<Utf8, Utf8>) valueRecord.get(STRING_MAP_FILED);
-            assertEquals(stringMapValue.get(new Utf8("k1")), new Utf8("v1"));
-          } catch (Exception e) {
-            throw new VeniceException(e);
-          }
-        });
-      }
-    }
-    sendStreamingRecord(veniceProducer, storeName, key1, updateBuilder.build(), 12345L);
-
-    for (int i = 0; i < NUMBER_OF_CHILD_DATACENTERS; i++) {
-      VeniceClusterWrapper cluster = childDatacenters.get(i).getClusters().get(CLUSTER_NAME);
-      try (AvroGenericStoreClient<Object, Object> storeReader = ClientFactory.getAndStartGenericAvroClient(
-          ClientConfig.defaultGenericClientConfig(storeName).setVeniceURL(cluster.getRandomRouterURL()))) {
-        TestUtils.waitForNonDeterministicAssertion(ASSERTION_TIMEOUT_MS, TimeUnit.MILLISECONDS, true, () -> {
-          try {
-            // Key 1 is preserved.
-            GenericRecord valueRecord = readValue(storeReader, key1);
-            assertNotNull(valueRecord);
-            assertNotNull(valueRecord.get(STRING_MAP_FILED));
-            Map<Utf8, Utf8> stringMapValue = (Map<Utf8, Utf8>) valueRecord.get(STRING_MAP_FILED);
-            assertEquals(stringMapValue.get(new Utf8("k1")), new Utf8("v1"));
-          } catch (Exception e) {
-            throw new VeniceException(e);
-          }
-        });
-      }
-    }
-  }
-
-  private void validateRmdData(
-      RmdSerDe rmdSerDe,
-      String kafkaTopic,
-      String key,
-      Consumer<RmdWithValueSchemaId> rmdDataValidationFlow) {
-    for (VeniceServerWrapper serverWrapper: multiRegionMultiClusterWrapper.getChildRegions()
-        .get(0)
-        .getClusters()
-        .get("venice-cluster0")
-        .getVeniceServers()) {
-      StorageEngine storageEngine = serverWrapper.getVeniceServer().getStorageService().getStorageEngine(kafkaTopic);
-      assertNotNull(storageEngine);
-      ValueRecord result = SingleGetChunkingAdapter
-          .getReplicationMetadata(storageEngine, 0, serializeStringKeyToByteArray(key), true, null);
-      // Avoid assertion failure logging massive RMD record.
-      boolean nullRmd = (result == null);
-      assertFalse(nullRmd);
-      byte[] value = result.serialize();
-      RmdWithValueSchemaId rmdWithValueSchemaId = new RmdWithValueSchemaId();
-      rmdSerDe.deserializeValueSchemaIdPrependedRmdBytes(value, rmdWithValueSchemaId);
-      rmdDataValidationFlow.accept(rmdWithValueSchemaId);
     }
   }
 
@@ -2086,234 +1350,6 @@ public class PartialUpdateTest {
         }
       });
 
-    } finally {
-      if (veniceProducer != null) {
-        veniceProducer.stop();
-      }
-    }
-  }
-
-  private GenericRecord readValue(AvroGenericStoreClient<Object, Object> storeReader, String key)
-      throws ExecutionException, InterruptedException {
-    return (GenericRecord) storeReader.get(key).get();
-  }
-
-  @Test(timeOut = TEST_TIMEOUT_MS, dataProvider = "Boolean-Compression", dataProviderClass = DataProviderUtils.class)
-  public void testWriteComputeWithHybridLeaderFollowerLargeRecord(
-      boolean writeComputeFromCache,
-      CompressionStrategy compressionStrategy) throws Exception {
-
-    SystemProducer veniceProducer = null;
-
-    try {
-      long streamingRewindSeconds = 10L;
-      long streamingMessageLag = 2L;
-
-      String storeName = Utils.getUniqueString("write-compute-store");
-      File inputDir = getTempDataDirectory();
-      String inputDirPath = "file://" + inputDir.getAbsolutePath();
-      String parentControllerURL = parentController.getControllerUrl();
-      // Records 1-100, id string to name record
-      Schema recordSchema = writeSimpleAvroFileWithStringToNameRecordV1Schema(inputDir);
-      VeniceClusterWrapper veniceClusterWrapper = childDatacenters.get(0).getClusters().get(CLUSTER_NAME);
-      Properties vpjProperties =
-          IntegrationTestPushUtils.defaultVPJProps(multiRegionMultiClusterWrapper, inputDirPath, storeName);
-      try (ControllerClient controllerClient = new ControllerClient(CLUSTER_NAME, parentControllerURL)) {
-
-        String keySchemaStr = recordSchema.getField(DEFAULT_KEY_FIELD_PROP).schema().toString();
-        String valueSchemaStr = recordSchema.getField(DEFAULT_VALUE_FIELD_PROP).schema().toString();
-        assertCommand(controllerClient.createNewStore(storeName, "test_owner", keySchemaStr, valueSchemaStr));
-
-        ControllerResponse response = controllerClient.updateStore(
-            storeName,
-            new UpdateStoreQueryParams().setHybridRewindSeconds(streamingRewindSeconds)
-                .setHybridOffsetLagThreshold(streamingMessageLag)
-                .setStorageQuotaInByte(Store.UNLIMITED_STORAGE_QUOTA)
-                .setChunkingEnabled(true)
-                .setCompressionStrategy(compressionStrategy)
-                .setWriteComputationEnabled(true)
-                .setHybridRewindSeconds(10L)
-                .setHybridOffsetLagThreshold(2L));
-
-        assertFalse(response.isError());
-
-        // Add a new value schema v2 to store
-        SchemaResponse schemaResponse = controllerClient.addValueSchema(storeName, NAME_RECORD_V2_SCHEMA.toString());
-        assertFalse(schemaResponse.isError());
-
-        // Add partial update schema associated to v2.
-        // Note that partial update schema needs to be registered manually here because the integration test harness
-        // does not create any parent controller. In production, when a value schema is added to a partial update
-        // enabled
-        // store via a parent controller, it will automatically generate and register its WC schema.
-        Schema writeComputeSchema =
-            WriteComputeSchemaConverter.getInstance().convertFromValueRecordSchema(NAME_RECORD_V2_SCHEMA);
-        schemaResponse =
-            controllerClient.addDerivedSchema(storeName, schemaResponse.getId(), writeComputeSchema.toString());
-        assertFalse(schemaResponse.isError());
-
-        // VPJ push
-        String childControllerUrl = childDatacenters.get(0).getRandomController().getControllerUrl();
-        try (ControllerClient childControllerClient = new ControllerClient(CLUSTER_NAME, childControllerUrl)) {
-          IntegrationTestPushUtils.runVPJ(vpjProperties, 1, childControllerClient);
-        }
-        veniceClusterWrapper.waitVersion(storeName, 1);
-        try (AvroGenericStoreClient<Object, Object> storeReader = ClientFactory.getAndStartGenericAvroClient(
-            ClientConfig.defaultGenericClientConfig(storeName)
-                .setVeniceURL(veniceClusterWrapper.getRandomRouterURL()))) {
-          // Verify records (note, records 1-100 have been pushed)
-          TestUtils.waitForNonDeterministicAssertion(10, TimeUnit.SECONDS, true, () -> {
-            try {
-              for (int i = 1; i < 100; i++) {
-                String key = String.valueOf(i);
-                GenericRecord value = readValue(storeReader, key);
-                assertNotNull(value, "Key " + key + " should not be missing!");
-                assertEquals(value.get("firstName").toString(), "first_name_" + key);
-                assertEquals(value.get("lastName").toString(), "last_name_" + key);
-                // We haven't written anything with the new schema yet
-                assertThrows(AvroRuntimeException.class, () -> value.get("age"));
-              }
-            } catch (Exception e) {
-              throw new VeniceException(e);
-            }
-          });
-
-          // disable the purging of transientRecord buffer using reflection.
-          if (writeComputeFromCache) {
-            String topicName = Version.composeKafkaTopic(storeName, 1);
-            for (VeniceServerWrapper veniceServerWrapper: veniceClusterWrapper.getVeniceServers()) {
-              StoreIngestionTaskBackdoor.setPurgeTransientRecordBuffer(veniceServerWrapper, topicName, false);
-            }
-          }
-
-          // Do not send large record to RT; RT doesn't support chunking
-          veniceProducer = getSamzaProducer(veniceClusterWrapper, storeName, Version.PushType.STREAM);
-          String key = String.valueOf(101);
-          GenericRecord value = new GenericData.Record(NAME_RECORD_V1_SCHEMA);
-          char[] chars = new char[100];
-          Arrays.fill(chars, 'f');
-          String firstName = new String(chars);
-          Arrays.fill(chars, 'l');
-          String lastName = new String(chars);
-          value.put("firstName", firstName);
-          value.put("lastName", lastName);
-          sendStreamingRecord(veniceProducer, storeName, key, value);
-
-          // Verify the streaming record
-          TestUtils.waitForNonDeterministicAssertion(60, TimeUnit.SECONDS, true, () -> {
-            try {
-              GenericRecord retrievedValue = readValue(storeReader, key);
-              assertNotNull(retrievedValue, "Key " + key + " should not be missing!");
-              assertEquals(retrievedValue.get("firstName").toString(), firstName);
-              assertEquals(retrievedValue.get("lastName").toString(), lastName);
-            } catch (Exception e) {
-              throw new VeniceException(e);
-            }
-          });
-
-          // Update the record
-          Arrays.fill(chars, 'u');
-          String updatedFirstName = new String(chars);
-          final int updatedAge = 1;
-          UpdateBuilder updateBuilder = new UpdateBuilderImpl(writeComputeSchema);
-          updateBuilder.setNewFieldValue("firstName", updatedFirstName);
-          updateBuilder.setNewFieldValue("age", updatedAge);
-          GenericRecord partialUpdateRecord = updateBuilder.build();
-
-          sendStreamingRecord(veniceProducer, storeName, key, partialUpdateRecord);
-          // Verify the update
-          TestUtils.waitForNonDeterministicAssertion(60, TimeUnit.SECONDS, true, () -> {
-            try {
-              GenericRecord retrievedValue = readValue(storeReader, key);
-              assertNotNull(retrievedValue, "Key " + key + " should not be missing!");
-              assertEquals(retrievedValue.get("firstName").toString(), updatedFirstName);
-              assertEquals(retrievedValue.get("lastName").toString(), lastName);
-              assertEquals(retrievedValue.get("age"), updatedAge);
-            } catch (Exception e) {
-              throw new VeniceException(e);
-            }
-          });
-
-          // Update the record again
-          Arrays.fill(chars, 'v');
-          String updatedFirstName1 = new String(chars);
-
-          updateBuilder = new UpdateBuilderImpl(writeComputeSchema);
-          updateBuilder.setNewFieldValue("firstName", updatedFirstName1);
-          GenericRecord partialUpdateRecord1 = updateBuilder.build();
-          sendStreamingRecord(veniceProducer, storeName, key, partialUpdateRecord1);
-          // Verify the update
-          TestUtils.waitForNonDeterministicAssertion(60, TimeUnit.SECONDS, () -> {
-            try {
-              GenericRecord retrievedValue = readValue(storeReader, key);
-              assertNotNull(retrievedValue, "Key " + key + " should not be missing!");
-              assertEquals(retrievedValue.get("firstName").toString(), updatedFirstName1);
-              assertEquals(retrievedValue.get("lastName").toString(), lastName);
-            } catch (Exception e) {
-              throw new VeniceException(e);
-            }
-          });
-
-          // Delete the record
-          sendStreamingRecord(veniceProducer, storeName, key, null);
-          // Verify the delete
-          TestUtils.waitForNonDeterministicAssertion(60, TimeUnit.SECONDS, true, () -> {
-            try {
-              GenericRecord retrievedValue = readValue(storeReader, key);
-              assertNull(retrievedValue, "Key " + key + " should be missing!");
-            } catch (Exception e) {
-              throw new VeniceException(e);
-            }
-          });
-
-          // Update the record again
-          Arrays.fill(chars, 'w');
-          String updatedFirstName2 = new String(chars);
-          Arrays.fill(chars, 'g');
-          String updatedLastName = new String(chars);
-
-          updateBuilder = new UpdateBuilderImpl(writeComputeSchema);
-          updateBuilder.setNewFieldValue("firstName", updatedFirstName2);
-          updateBuilder.setNewFieldValue("lastName", updatedLastName);
-          updateBuilder.setNewFieldValue("age", 2);
-          GenericRecord partialUpdateRecord2 = updateBuilder.build();
-
-          sendStreamingRecord(veniceProducer, storeName, key, partialUpdateRecord2);
-          // Verify the update
-          TestUtils.waitForNonDeterministicAssertion(60, TimeUnit.SECONDS, true, () -> {
-            try {
-              GenericRecord retrievedValue = readValue(storeReader, key);
-              assertNotNull(retrievedValue, "Key " + key + " should not be missing!");
-              assertEquals(retrievedValue.get("firstName").toString(), updatedFirstName2);
-              assertEquals(retrievedValue.get("lastName").toString(), updatedLastName);
-              assertEquals(retrievedValue.get("age"), 2);
-            } catch (Exception e) {
-              throw new VeniceException(e);
-            }
-          });
-
-          // Update the record again
-          Arrays.fill(chars, 'x');
-          String updatedFirstName3 = new String(chars);
-
-          updateBuilder = new UpdateBuilderImpl(writeComputeSchema);
-          updateBuilder.setNewFieldValue("firstName", updatedFirstName3);
-          GenericRecord partialUpdateRecord3 = updateBuilder.build();
-          sendStreamingRecord(veniceProducer, storeName, key, partialUpdateRecord3);
-          // Verify the update
-          TestUtils.waitForNonDeterministicAssertion(60, TimeUnit.SECONDS, true, () -> {
-            try {
-              GenericRecord retrievedValue = readValue(storeReader, key);
-              assertNotNull(retrievedValue, "Key " + key + " should not be missing!");
-              assertEquals(retrievedValue.get("firstName").toString(), updatedFirstName3);
-              assertEquals(retrievedValue.get("lastName").toString(), updatedLastName);
-            } catch (Exception e) {
-              throw new VeniceException(e);
-            }
-          });
-
-        }
-      }
     } finally {
       if (veniceProducer != null) {
         veniceProducer.stop();
@@ -2544,4 +1580,32 @@ public class PartialUpdateTest {
     }
   }
 
+  private void validateRmdData(
+      RmdSerDe rmdSerDe,
+      String kafkaTopic,
+      String key,
+      Consumer<RmdWithValueSchemaId> rmdDataValidationFlow) {
+    for (VeniceServerWrapper serverWrapper: multiRegionMultiClusterWrapper.getChildRegions()
+        .get(0)
+        .getClusters()
+        .get("venice-cluster0")
+        .getVeniceServers()) {
+      StorageEngine storageEngine = serverWrapper.getVeniceServer().getStorageService().getStorageEngine(kafkaTopic);
+      assertNotNull(storageEngine);
+      ValueRecord result = SingleGetChunkingAdapter
+          .getReplicationMetadata(storageEngine, 0, serializeStringKeyToByteArray(key), true, null);
+      // Avoid assertion failure logging massive RMD record.
+      boolean nullRmd = (result == null);
+      assertFalse(nullRmd);
+      byte[] value = result.serialize();
+      RmdWithValueSchemaId rmdWithValueSchemaId = new RmdWithValueSchemaId();
+      rmdSerDe.deserializeValueSchemaIdPrependedRmdBytes(value, rmdWithValueSchemaId);
+      rmdDataValidationFlow.accept(rmdWithValueSchemaId);
+    }
+  }
+
+  private GenericRecord readValue(AvroGenericStoreClient<Object, Object> storeReader, String key)
+      throws ExecutionException, InterruptedException {
+    return (GenericRecord) storeReader.get(key).get();
+  }
 }
