--- internal/venice-test-common/src/integrationTest/java/com/linkedin/venice/endToEnd/PartialUpdateWithParallelProcessingTest.java
@@ -1,8 +1,0 @@
-package com.linkedin.venice.endToEnd;
-
-public class PartialUpdateWithParallelProcessingTest extends PartialUpdateTest {
-  @Override
-  protected boolean isAAWCParallelProcessingEnabled() {
-    return true;
-  }
-}
